requirements.txt
chromadb
openai
python-dotenv
watchdog
tomli; python_version < "3.11"
